refactor: step5 리팩토링
