+++ src/main/kotlin/carracing/view/ResultView.kt
@@ -1,4 +1,4 @@
-package carracing.ui
+package carracing.view
 
 import carracing.domain.Car
